fix: correct naming of routes
