+++ apps/web/src/app/api/(routes)/sanity/draft-mode/disable/route.ts
@@ -1,12 +1,8 @@
-/**
- * This file is used to allow Presentation to set the app in Draft Mode, which will load Visual Editing
- * and query draft content and preview the content as it will appear once everything is published
- */
+import { draftMode } from 'next/headers';
+import { NextRequest, NextResponse } from 'next/server';
 
-import { defineEnableDraftMode } from 'next-sanity/draft-mode';
-import { client } from '@/features/sanity/data/client';
-import { appConfig } from '@/config/app';
+export async function GET(request: NextRequest) {
+  (await draftMode()).disable();
 
-export const { GET } = defineEnableDraftMode({
-  client: client.withConfig({ token: appConfig.sanity.readToken }),
-});
+  return NextResponse.redirect(new URL('/', request.url));
+}
